+++ ecs-init/config/release.go
@@ -1,6 +1,3 @@
-//go:build !development
-// +build !development
-
 // Copyright 2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may