+++ ecs-init/engine/engine.go
@@ -21,6 +21,7 @@ import (
 	"os"
 	"time"
 
+	"github.com/aws/amazon-ecs-agent/ecs-init/apparmor"
 	"github.com/aws/amazon-ecs-agent/ecs-init/backoff"
 	"github.com/aws/amazon-ecs-agent/ecs-init/cache"
 	"github.com/aws/amazon-ecs-agent/ecs-init/config"
@@ -31,6 +32,7 @@ import (
 	"github.com/aws/amazon-ecs-agent/ecs-init/gpu"
 
 	log "github.com/cihub/seelog"
+	ctrdapparmor "github.com/containerd/containerd/pkg/apparmor"
 )
 
 const (
@@ -49,9 +51,13 @@ const (
 )
 
 // Injection point for testing purposes
-var getDockerClient = func() (dockerClient, error) {
-	return docker.Client()
-}
+var (
+	getDockerClient = func() (dockerClient, error) {
+		return docker.Client()
+	}
+	hostSupports       = ctrdapparmor.HostSupports
+	loadDefaultProfile = apparmor.LoadDefaultProfile
+)
 
 func dockerError(err error) error {
 	return engineError("could not create docker client", err)
@@ -110,6 +116,11 @@ func New() (*Engine, error) {
 func (e *Engine) PreStart() error {
 	// setup gpu if necessary
 	err := e.PreStartGPU()
+	if err != nil {
+		return err
+	}
+	// setup AppArmor if necessary
+	err = e.PreStartAppArmor()
 	if err != nil {
 		return err
 	}
@@ -191,6 +202,16 @@ func (e *Engine) PreStartGPU() error {
 				return engineError("Nvidia GPU Manager", err)
 			}
 		}
+	}
+	return nil
+}
+
+// PreStartAppArmor sets up the ecs-default AppArmor profile if we're running
+// on an AppArmor-enabled system.
+func (e *Engine) PreStartAppArmor() error {
+	if hostSupports() {
+		log.Infof("pre-start: setting up %s AppArmor profile", apparmor.ECSDefaultProfileName)
+		return loadDefaultProfile(apparmor.ECSDefaultProfileName)
 	}
 	return nil
 }